refactor: rename dbt sqlbuild work bridge
