+++ src/sqlbuild/cli/commands/main/dbt_sqlbuild_work.py
@@ -1,4 +1,4 @@
-"""SQLBuild execution public entry for dbt interop orchestration."""
+"""SQLBuild work public entry for dbt interop orchestration."""
 
 from __future__ import annotations
 
@@ -6,7 +6,7 @@ from pathlib import Path
 from typing import TextIO
 
 from sqlbuild.adapter.base.base_adapter import BaseAdapter
-from sqlbuild.cli.commands.main.helpers.dbt_sqlbuild_execution import (
+from sqlbuild.cli.commands.main.helpers.dbt_sqlbuild_work import (
     execute_sqlbuild_build_work,
     execute_sqlbuild_test_work,
 )
