+++ src/sqlbuild/integrations/dbt/pipeline/main/execute.py
@@ -11,7 +11,7 @@ from typing import Any, TextIO
 
 from sqlbuild.adapter.base.base_adapter import BaseAdapter
 from sqlbuild.cli.commands.main.connection_progress import build_connection_progress_reporter
-from sqlbuild.cli.commands.main.dbt_sqlbuild_execution import execute_dbt_sqlbuild_work
+from sqlbuild.cli.commands.main.dbt_sqlbuild_work import execute_dbt_sqlbuild_work
 from sqlbuild.compiler.compile.main.effective_config import build_effective_connection_config
 from sqlbuild.compiler.compile.models import CompiledProject
 from sqlbuild.compiler.discovery.main.discover import discover_project_inputs
